Chore: Moving functions into common
